Create main.cpp
Initializes the string variable with the given value. It has its own memory adress. Creation of pointer which points to the string value but saves it on its own adress. Reference to the first memory adress, so its just another name for it. Then we print the memory addresses and then the values