ex06 Create main.cpp
getLevelIndex to convert the DEBUG INFO WARNING or ERROR to numbers so i can use it in the switch statement. Int Main checks if there are 2 arguments if not there is a usage message. If so it checks which level it is so DEBUG INFO ... and if not it says the default message